+++ app/src/main/java/com/example/king_bob_nae/features/mykkilog/presentation/MySimpleKkiLogFragment.kt
@@ -1,15 +1,15 @@
-package com.example.king_bob_nae.features.mykkilog
+package com.example.king_bob_nae.features.mykkilog.presentation
 
 import android.os.Bundle
 import android.view.View
 import androidx.recyclerview.widget.GridLayoutManager
 import com.example.king_bob_nae.R
 import com.example.king_bob_nae.base.BaseFragment
-import com.example.king_bob_nae.databinding.FragmentMyDetailKkilogBinding
+import com.example.king_bob_nae.databinding.FragmentMySimpleKkilogBinding
 import com.example.king_bob_nae.features.mykkilog.data.MyKkiLog
 
-class MyDetailKkiLogFragment :
-    BaseFragment<FragmentMyDetailKkilogBinding>(R.layout.fragment_my_detail_kkilog) {
+class MySimpleKkiLogFragment :
+    BaseFragment<FragmentMySimpleKkilogBinding>(R.layout.fragment_my_simple_kkilog) {
     private val itemAdapter: MyKkiLogAdapter by lazy { MyKkiLogAdapter() }
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -24,7 +24,7 @@ class MyDetailKkiLogFragment :
         val item5 = MyKkiLog(title = "존맛탱 김치", date = "2022.08.24")
         val item6 = MyKkiLog(title = "썩은 김치", date = "2022.08.24")
         val itemList: List<MyKkiLog> = listOf(item1, item2, item3, item4, item5, item6)
-        binding.rvMyDetailKkilog.apply {
+        binding.rvMySimpleKkilog.apply {
             layoutManager = GridLayoutManager(requireContext(), 2)
             adapter = itemAdapter
         }
